Extract species number of any plant family

Produces a CSV file listing the number of species and genera within the target botanical families of flowering plants available at Plants of the World Online (POWO).

Usage

powoFam(family,
        verbose = TRUE,
        save = FALSE,
        dir = "results_powoFam",
        filename = "output")

Arguments

family

Either one family name or a vector of multiple families that is present in POWO.

verbose

Logical, if FALSE, a message showing each step during the POWO search will not be printed in the console in full.

save

Logical, if TRUE, the search results will be saved on disk.

dir

Pathway to the computer's directory, where the file will be saved provided that the argument save is set up in TRUE. The default is to create a directory named results_powoFam and the searched results will be saved within a subfolder named after the current date.

filename

Name of the output file to be saved. The default is to create a file entitled output.

Value

A table with the following fields: family, genus_number (counting of genera within the specified plant family), species_number (countig of species within each genus), kew_id (the family code used in Kew's database), powo_uri (the URI to access the family in POWO).

See also

Author

Debora Zuanny & Domingos Cardoso

Examples

if (FALSE) {
library(expowo)

powoFam(family = "Lecythidaceae",
        save = FALSE,
        dir = "results_powoFam",
        filename = "Lecythidaceae_spp_number")
}