Extract list of species from POWO

Produces a CSV file listing all accepted species and associated geographical distribution from any target genus or family of flowering plants at Plants of the World Online (POWO).

Usage

powoSpecies(family, genus = NULL, hybridspp = FALSE, country = NULL,
            verbose = TRUE, save = FALSE, dir, filename)

Arguments

family

Either one family name or a vector of multiple families that is present in POWO.

genus

Either one genus name or a vector of multiple genera that are present in POWO. If any genus name is not provided, then the function will search any species from all accepted genera known for the target family.

hybridspp

Logical, if TRUE, the search results will include hybrid species.

country

Either one country name or a vector of multiple countries. If country names are provided, then the function will return only the species that are native to such countries, according to POWO.

verbose

Logical, if FALSE, the search results will not be printed in the console in full.

save

Logical, if FALSE, the search results will not be saved on disk.

dir

Pathway to the computer's directory, where the file will be saved provided that the argument save is set up in TRUE. The default is to create a directory named results_powoSpecies and the search results will be saved within a subfolder named by the current date.

filename

Name of the output file to be saved. The default is to create a file entitled output.

Value

Table in .csv format.

See also

Author

Debora Zuanny & Domingos Cardoso

Examples

# \donttest{
library(expowo)

powoSpecies(family = "Martyniaceae",
            hybridspp = FALSE,
            country = c("Argentina", "Brazil", "French Guiana"),
            verbose = TRUE,
            save = FALSE,
            dir = "results_powoSpecies/",
            filename = "Martyniaceae_spp")
#> [1] "Searching... Martyniaceae 1/1"
#> [1] "Searching spp list of... Craniolaria Martyniaceae 1/5"
#> [1] "Searching spp list of... Holoregmia Martyniaceae 2/5"
#> [1] "Searching spp list of... Ibicella Martyniaceae 3/5"
#> [1] "Searching spp list of... Martynia Martyniaceae 4/5"
#> [1] "Searching spp list of... Proboscidea Martyniaceae 5/5"
#> [1] "Searching distribution of... Craniolaria annua Martyniaceae 1/14"
#> [1] "Searching distribution of... Craniolaria integrifolia Martyniaceae 2/14"
#> [1] "Searching distribution of... Holoregmia viscida Martyniaceae 3/14"
#> [1] "Searching distribution of... Ibicella lutea Martyniaceae 4/14"
#> [1] "Searching distribution of... Ibicella parodii Martyniaceae 5/14"
#> [1] "Searching distribution of... Martynia annua Martyniaceae 6/14"
#> [1] "Searching distribution of... Martynia palmeri Martyniaceae 7/14"
#> [1] "Searching distribution of... Proboscidea althaeifolia Martyniaceae 8/14"
#> [1] "Searching distribution of... Proboscidea diversifolia Martyniaceae 9/14"
#> [1] "Searching distribution of... Proboscidea louisianica Martyniaceae 10/14"
#> [1] "Searching distribution of... Proboscidea parviflora Martyniaceae 11/14"
#> [1] "Searching distribution of... Proboscidea sabulosa Martyniaceae 12/14"
#> [1] "Searching distribution of... Proboscidea spicata Martyniaceae 13/14"
#> [1] "Searching distribution of... Proboscidea triloba Martyniaceae 14/14"
#> Your search returned species with distribution only in the
#>                     following countries:
#>  
#>  Argentina, Brazil 
#>  
#>  There is no species occurring in the countries below:
#>  
#>  French Guiana 
#>  
#>  Check whether any species does not occur in the countries
#>                     above either because:
#>  1. The species indeed does not occur in the provided
#>                     country vector;
#>  2. The country name is written with any typo;
#>  3. Any country name in the country vector is not written
#>                     in English language.
#>  
#> Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Craniolaria Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Martynia Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Martynia Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Proboscidea Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Proboscidea Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Proboscidea Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Proboscidea Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Proboscidea Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Proboscidea Genera listed below were removed from the original search
#>                   because they are not native to any of the given country
#>                   vector:
#>  
#>  Proboscidea
#>         family       genus      species               taxon_name
#> 2 Martyniaceae Craniolaria integrifolia Craniolaria integrifolia
#> 3 Martyniaceae  Holoregmia      viscida       Holoregmia viscida
#> 4 Martyniaceae    Ibicella        lutea           Ibicella lutea
#> 5 Martyniaceae    Ibicella      parodii         Ibicella parodii
#>               authors                    scientific_name
#> 2               Cham.     Craniolaria integrifolia Cham.
#> 3                Nees            Holoregmia viscida Nees
#> 4 (Lindl.) Van Eselt. Ibicella lutea (Lindl.) Van Eselt.
#> 5            Abbiatti          Ibicella parodii Abbiatti
#>                                                     publication
#> 2                                         Linnaea 7: 725 (1832)
#> 3                                           Flora 4: 300 (1821)
#> 4 Techn. Bull. New York Agric. Exp. Sta., Geneva 149: 34 (1929)
#> 5                       Notas Mus. La Plata, Bot. 4: 458 (1939)
#>                               native_to_country
#> 2 Argentina, Bolivia, Brazil, Paraguay, Uruguay
#> 3                                        Brazil
#> 4          Argentina, Brazil, Paraguay, Uruguay
#> 5                            Argentina, Bolivia
#>                                                                                                                 native_to_botanical_countries
#> 2 Argentina Northeast, Argentina Northwest, Bolivia, Brazil Northeast, Brazil South, Brazil Southeast, Brazil West-Central, Paraguay, Uruguay
#> 3                                                                                                                            Brazil Northeast
#> 4                               Argentina Northeast, Argentina Northwest, Brazil Northeast, Brazil South, Brazil Southeast, Paraguay, Uruguay
#> 5                                                                                           Argentina Northeast, Argentina Northwest, Bolivia
#>                                                             introduced_to_country
#> 2                                                                         unknown
#> 3                                                                         unknown
#> 4 Algeria, Belgium, France, Germany, Greece, South Africa, Tunisia, United States
#> 5                                                                         unknown
#>                                                                                                                           introduced_to_botanical_countries
#> 2                                                                                                                                                   unknown
#> 3                                                                                                                                                   unknown
#> 4 Algeria, Arizona, Belgium, California, Cape Provinces, East Aegean Is., Florida, France, Germany, KwaZulu-Natal, Mississippi, Northern Provinces, Tunisia
#> 5                                                                                                                                                   unknown
#>     kew_id
#> 2 675805-1
#> 3 327210-2
#> 4 565368-1
#> 5 127285-2
#>                                                                       powo_uri
#> 2 http://www.plantsoftheworldonline.org/taxon/urn:lsid:ipni.org:names:675805-1
#> 3 http://www.plantsoftheworldonline.org/taxon/urn:lsid:ipni.org:names:327210-2
#> 4 http://www.plantsoftheworldonline.org/taxon/urn:lsid:ipni.org:names:565368-1
#> 5 http://www.plantsoftheworldonline.org/taxon/urn:lsid:ipni.org:names:127285-2
# }