Create global maps of species richness

It produces global-scale maps of species richness at country and botanical country levels for any genus or family of flowering plants with associated species distribution data at Plants of the World Online (POWO).

Usage

powoMap(data = NULL,
        botctrs = NULL,
        map_div = "native_to_country",
        multigen = FALSE,
        verbose = TRUE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_map",
        format = "jpg")

Arguments

data

A data frame of species' global distribution at country or botanical country level as created with the function powoSpecies.

botctrs

An object of class sf data.frame with botanical country division of the World. This object comes as the expowo's data package botdivmap.

map_div

One of the two column names of native distribution from the input data either created with powoSpecies or from an external spreadsheet. The default of this argument is "native_to_country" to generate global species richness maps at country level. By choosing "native_to_botanical_countries" the global species richness maps will be colored according to the level 3 of botanical subdivision of the World Geographical Scheme for Recording Plant Distributions. But then make sure to provided in the argument botctrs the World's botanical country divisions from botdivmap. If the input is generated by external sources, the data and columns of the spreadsheet must match with the standardized output from powoSpecies.

multigen

Logical, if TRUE, the function will produce multiple global maps for every single genus within the input data.

verbose

Logical, if FALSE, the map creation steps will not be printed in the console in full.

vir_color

A character vector with the name or code of any of the color palettes from Viridis package.

bre_color

A character vector with the name or code of any of the color palettes from RColorBrewer package.

leg_title

A character vector to be displayed in the output map as a legend. Default is to create one title called SR, which means species richness.

dpi

One number in the range of 72-4000 referring to the image resolution in the format of dots per inch in the output file. Default is to create an output with 600 dpi.

dir

Pathway to the computer's directory, where the map file will be saved provided that the argument save is set up in TRUE. The default is to create a directory named results_powoMap/ and the search results will be saved within a subfolder named by the current date.

filename

Name of the output file to be saved. Default is to create a file entitled global_richness_map.

format

A character vector related to the file format of the global map to be saved. The default is "jpg" to save the output in Joint Photographic Experts Group (.jpg), but you can also choose "pdf" to save in Portable Document Format (.pdf), "tiff" to save in Tag Image File Format (.tiff) or "png" to save in Portable Network Graphics (.png).

Value

Global map in .pdf format and saves the output on disk.

See also

Author

Debora Zuanny & Domingos Cardoso

Examples

if (FALSE) {
library(expowo)
library(taxize)

powocodes <- cbind(family = "Lecythidaceae",
                   data.frame(taxize::get_pow("Lecythidaceae")))

mapspdist <- powoSpecies(powocodes$family, powocodes$uri,
                         hybridspp = FALSE,
                         country = NULL,
                         verbose = TRUE,
                         save = TRUE,
                         dir = "results_powoSpecies/",
                         filename = "Lecythidaceae")

# To create a map according to political countries
powoMap(data = mapspdist,
        botctrs = NULL,
        map_div = "native_to_country",
        multigen = TRUE,
        verbose = FALSE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_country_map",
        format = "jpg")

# To create a map according to botanical countries subdivision
data(botdivmap)
powoMap(data = mapspdist,
        botctrs = botdivmap,
        map_div = "native_to_botanical_countries",
        multigen = TRUE,
        verbose = FALSE,
        vir_color = "viridis",
        bre_color = NULL,
        leg_title = "SR",
        dpi = 600,
        dir = "results_powoMap/",
        filename = "global_richness_botcountry_map",
        format = "jpg")
}