Get list of species from POWO

It produces a CSV file listing all accepted species and associated geographical distribution from any target genus or family of flowering plants at Plants of the World Online (POWO).

powoSpecies(family, uri, genus = NULL, hybridspp = FALSE, country = NULL,
            verbose = TRUE, save = TRUE, dir, filename)

Arguments

family

Either a single family name or a vector of multiple families that are present in POWO.

uri

URI address for each family associated to the target genus to be searched in POWO.

genus

Either one genus name or a vector of multiple genera that are present in POWO. If any genus name is not provided, then the function will search any species from all accepted genera known for the target family.

hybridspp

Logical, if TRUE, the search results will include hybrid species.

country

Either one country name or a vector of multiple countries. If country names are provided, then the function will return only the species that are native to such countries, according to POWO.

verbose

Logical, if FALSE, the searched results will not be printed in the console in full.

save

Logical, if FALSE, the searched results will not be saved on disk.

dir

Pathway to the computer's directory, where the file will be saved provided that the argument save is set up in TRUE. The default is to create a directory named results_powoSpecies and the searched results will be saved within a subfolder named by the current date.

filename

Name of the output file to be saved. Default is to create a file entitled output.

Value

Table in .csv format and saves the output on disk.

See also

Author

Debora Zuanny & Domingos Cardoso

Examples

if (FALSE) {
library(expowo)
library(taxize)

fam <- c("Araceae", "Lecythidaceae")
powocodes <- cbind(family = fam,
                   data.frame(taxize::get_pow(fam)))

powoSpecies(powocodes$family, powocodes$uri,
            hybridspp = FALSE,
            country = c("Argentina", "Brazil", "French Guiana"),
            verbose = TRUE,
            save = TRUE,
            dir = "results_powoSpecies/",
            filename = "Araceae_Lecythidaceae")

## Searching for all species and associated global distribution
## in any or all flowering plant genera, by using the URI addresses
## within the POWOcodes data file

data(POWOcodes)

powoSpecies(powocodes$family, powocodes$uri,
            hybridspp = TRUE,
            verbose = TRUE,
            save = TRUE,
            dir = "results_powoSpecies/",
            filename = "all_angiosperm_species")
}